Initial commit
Generate a simple HTML document representing the hierarchy of 
file and folders in the given directory. Dirdiv is a perfect
tool for creating a read-only representation of a directory 
structure for online tutorials or print media. Folder and file
icons are included by default, along with just the right amount of
styling.

Usage: python dirdiv.py c:\projects\demo